==> voice_of_customer/src/2_sampling_merged_output.ipynb <==
**Loading Requirements**

import pandas as pd
import datetime

import os
path = r'/N/u/abbajpai/Carbonate/Documents/Social_Media_Mining/data/output'

**Reading** **Merged** **Output** **File**

# read the merged output from previous step
mergedData = pd.read_pickle(os.path.join(path,'2_mergedData.pkl'))

mergedData['date'] = pd.to_datetime(mergedData['reviewTime']) # review date from UNIX format

mergedData['lenReview']= mergedData['reviewText'].str.split().str.len() # finding number of words in each review

len(mergedData)

**Sampling** **Data**

# 1. Filter on length of review (minimum 10 words)
mergedDataSubset = mergedData.query("lenReview > 10")

# 2. Filter on time duration (year greater than 2015)
TopBrands = mergedDataSubset[mergedDataSubset.date.dt.year>=2015]

# 3. Drop all duplicate reviews
TopBrands = TopBrands.drop_duplicates(ignore_index=True, keep='first', subset='reviewText')

**Finding Top Brands in Filtered Data**

TopBrands.groupby(TopBrands.date.dt.year)['asin'].count().sort_values(ascending=False)

len(TopBrands)

brands = (TopBrands.groupby('brand')['asin'].count().sort_values(ascending=False).head(5)).index.tolist()

Top5Brands = TopBrands[TopBrands['brand'].isin(brands)]
print( len(Top5Brands))

**Restricting Sample Size to 100K**

Top5BrandSample = Top5Brands.sample(n=100000, random_state=1)
print(len(Top5BrandSample))

' Other protocols have issues while reading on Colab'
Top5BrandSample.to_pickle(os.path.join(path,'Top5BrandSample.pkl'),protocol =4)
